+++ src/quickpicks/remoteProviderPicker.ts
@@ -1,5 +1,5 @@
 'use strict';
-import { Disposable, env, Uri, window } from 'vscode';
+import { Disposable, env, QuickInputButton, ThemeIcon, Uri, window } from 'vscode';
 import { Commands, OpenOnRemoteCommandArgs } from '../commands';
 import { GlyphChars } from '../constants';
 import { Container } from '../container';
@@ -31,10 +31,12 @@ export class CopyOrOpenRemoteCommandQuickPickItem extends CommandQuickPickItem {
 		private readonly remote: GitRemote<RemoteProvider>,
 		private readonly resource: RemoteResource,
 		private readonly clipboard?: boolean,
+		buttons?: QuickInputButton[],
 	) {
 		super({
 			label: `$(repo) ${remote.provider.path}`,
 			description: remote.name,
+			buttons: buttons,
 		});
 	}
 
@@ -86,6 +88,10 @@ export class CopyOrOpenRemoteCommandQuickPickItem extends CommandQuickPickItem {
 
 		void (await (this.clipboard ? this.remote.provider.copy(resource) : this.remote.provider.open(resource)));
 	}
+
+	setAsDefault(): Promise<void> {
+		return this.remote.setAsDefault(true);
+	}
 }
 
 export class CopyRemoteResourceCommandQuickPickItem extends CommandQuickPickItem {
@@ -131,28 +137,11 @@ export class OpenRemoteResourceCommandQuickPickItem extends CommandQuickPickItem
 	}
 }
 
-export class SetADefaultRemoteCommandQuickPickItem extends CommandQuickPickItem {
-	constructor(private readonly remotes: GitRemote<RemoteProvider>[]) {
-		super({ label: 'Set a Default Remote...' });
-	}
-
-	override async execute(): Promise<GitRemote<RemoteProvider> | undefined> {
-		return RemoteProviderPicker.setADefaultRemote(this.remotes);
-	}
-}
-
-export class SetRemoteAsDefaultCommandQuickPickItem extends CommandQuickPickItem {
-	constructor(private readonly remote: GitRemote<RemoteProvider>) {
-		super({
-			label: remote.provider.name,
-			detail: `$(repo) ${remote.provider.path}`,
-		});
-	}
-
-	override async execute(): Promise<GitRemote<RemoteProvider>> {
-		void (await this.remote.setAsDefault(true));
-		return this.remote;
-	}
+namespace QuickCommandButtons {
+	export const SetRemoteAsDefault: QuickInputButton = {
+		iconPath: new ThemeIcon('settings-gear'),
+		tooltip: 'Set as Default Remote',
+	};
 }
 
 export namespace RemoteProviderPicker {
@@ -162,19 +151,10 @@ export namespace RemoteProviderPicker {
 		resource: RemoteResource,
 		remotes: GitRemote<RemoteProvider>[],
 		options?: { autoPick?: 'default' | boolean; clipboard?: boolean; setDefault?: boolean },
-	): Promise<
-		| ConfigureCustomRemoteProviderCommandQuickPickItem
-		| CopyOrOpenRemoteCommandQuickPickItem
-		| SetADefaultRemoteCommandQuickPickItem
-		| undefined
-	> {
+	): Promise<ConfigureCustomRemoteProviderCommandQuickPickItem | CopyOrOpenRemoteCommandQuickPickItem | undefined> {
 		const { autoPick, clipboard, setDefault } = { autoPick: false, clipboard: false, setDefault: true, ...options };
 
-		let items: (
-			| ConfigureCustomRemoteProviderCommandQuickPickItem
-			| CopyOrOpenRemoteCommandQuickPickItem
-			| SetADefaultRemoteCommandQuickPickItem
-		)[];
+		let items: (ConfigureCustomRemoteProviderCommandQuickPickItem | CopyOrOpenRemoteCommandQuickPickItem)[];
 		if (remotes.length === 0) {
 			items = [new ConfigureCustomRemoteProviderCommandQuickPickItem()];
 			placeHolder = 'No auto-detected or configured remote providers found';
@@ -187,36 +167,46 @@ export namespace RemoteProviderPicker {
 				}
 			}
 
-			items = remotes.map(r => new CopyOrOpenRemoteCommandQuickPickItem(r, resource, clipboard));
-			if (setDefault) {
-				items.push(new SetADefaultRemoteCommandQuickPickItem(remotes));
-			}
+			items = remotes.map(
+				r =>
+					new CopyOrOpenRemoteCommandQuickPickItem(
+						r,
+						resource,
+						clipboard,
+						setDefault ? [QuickCommandButtons.SetRemoteAsDefault] : undefined,
+					),
+			);
 		}
 
 		// eslint-disable-next-line @typescript-eslint/strict-boolean-expressions
 		if (autoPick && remotes.length === 1) return items[0];
 
 		const quickpick = window.createQuickPick<
-			| ConfigureCustomRemoteProviderCommandQuickPickItem
-			| CopyOrOpenRemoteCommandQuickPickItem
-			| SetADefaultRemoteCommandQuickPickItem
+			ConfigureCustomRemoteProviderCommandQuickPickItem | CopyOrOpenRemoteCommandQuickPickItem
 		>();
+		(quickpick as any).enableProposedApi = true;
 		quickpick.ignoreFocusOut = getQuickPickIgnoreFocusOut();
 
 		const disposables: Disposable[] = [];
 
 		try {
 			const pick = await new Promise<
-				| ConfigureCustomRemoteProviderCommandQuickPickItem
-				| CopyOrOpenRemoteCommandQuickPickItem
-				| SetADefaultRemoteCommandQuickPickItem
-				| undefined
+				ConfigureCustomRemoteProviderCommandQuickPickItem | CopyOrOpenRemoteCommandQuickPickItem | undefined
 			>(resolve => {
 				disposables.push(
 					quickpick.onDidHide(() => resolve(undefined)),
 					quickpick.onDidAccept(() => {
 						if (quickpick.activeItems.length !== 0) {
 							resolve(quickpick.activeItems[0]);
+						}
+					}),
+					quickpick.onDidTriggerItemButton(async e => {
+						if (
+							e.button === QuickCommandButtons.SetRemoteAsDefault &&
+							e.item instanceof CopyOrOpenRemoteCommandQuickPickItem
+						) {
+							void (await e.item.setAsDefault());
+							resolve(e.item);
 						}
 					}),
 				);
@@ -231,43 +221,6 @@ export namespace RemoteProviderPicker {
 			if (pick == null) return undefined;
 
 			return pick;
-		} finally {
-			quickpick.dispose();
-			disposables.forEach(d => d.dispose());
-		}
-	}
-
-	export async function setADefaultRemote(
-		remotes: GitRemote<RemoteProvider>[],
-	): Promise<GitRemote<RemoteProvider> | undefined> {
-		const items = remotes.map(r => new SetRemoteAsDefaultCommandQuickPickItem(r));
-
-		const quickpick = window.createQuickPick<SetRemoteAsDefaultCommandQuickPickItem>();
-		quickpick.ignoreFocusOut = getQuickPickIgnoreFocusOut();
-
-		const disposables: Disposable[] = [];
-
-		try {
-			const pick = await new Promise<SetRemoteAsDefaultCommandQuickPickItem | undefined>(resolve => {
-				disposables.push(
-					quickpick.onDidHide(() => resolve(undefined)),
-					quickpick.onDidAccept(() => {
-						if (quickpick.activeItems.length !== 0) {
-							resolve(quickpick.activeItems[0]);
-						}
-					}),
-				);
-
-				quickpick.title = 'Set a Default Remote';
-				quickpick.placeholder = 'Choose which remote to set as the default';
-				quickpick.matchOnDetail = true;
-				quickpick.items = items;
-
-				quickpick.show();
-			});
-			if (pick == null) return undefined;
-
-			return await pick.execute();
 		} finally {
 			quickpick.dispose();
 			disposables.forEach(d => d.dispose());